+++ weblate/utils/search.py
@@ -32,7 +32,12 @@ from jellyfish import damerau_levenshtein_distance
 from whoosh.fields import BOOLEAN, DATETIME, NUMERIC, TEXT, Schema
 
 from weblate.trans.util import PLURAL_SEPARATOR
-from weblate.utils.state import STATE_NAMES, STATE_TRANSLATED
+from weblate.utils.state import (
+    STATE_APPROVED,
+    STATE_NAMES,
+    STATE_READONLY,
+    STATE_TRANSLATED,
+)
 
 
 class Comparer:
@@ -112,34 +117,34 @@ class QueryParser(whoosh.qparser.QueryParser):
 
     def __init__(self):
         # Define fields for parsing
-        schema = Schema(
+        fields = {
             # Unit fields
-            source=TEXT,
-            target=TEXT,
-            context=TEXT,
-            note=TEXT,
-            location=TEXT,
-            priority=NumberField,
-            added=DATETIME,
-            state=StateField,
-            pending=BOOLEAN,
-            has=TEXT,
+            "source": TEXT,
+            "target": TEXT,
+            "context": TEXT,
+            "note": TEXT,
+            "location": TEXT,
+            "priority": NumberField,
+            "added": DATETIME,
+            "state": StateField,
+            "pending": BOOLEAN,
+            "has": TEXT,
+            "is": TEXT,
             # Language
-            language=TEXT,
+            "language": TEXT,
             # Change fields
-            changed=DATETIME,
-            changed_by=TEXT,
+            "changed": DATETIME,
+            "changed_by": TEXT,
             # Unit data
-            check=TEXT,
-            ignored_check=TEXT,
-            suggestion=TEXT,
-            suggestion_author=TEXT,
-            comment=TEXT,
-            comment_author=TEXT,
-            label=TEXT,
-        )
-        # Features to implement and corresponding blockers
-        # - unitdata lookups, https://github.com/WeblateOrg/weblate/issues/3007
+            "check": TEXT,
+            "ignored_check": TEXT,
+            "suggestion": TEXT,
+            "suggestion_author": TEXT,
+            "comment": TEXT,
+            "comment_author": TEXT,
+            "label": TEXT,
+        }
+        schema = Schema(**fields)
 
         # List of plugins
         plugins = [
@@ -245,6 +250,21 @@ def has_sql(text):
     raise ValueError("Unsupported has lookup: {}".format(text))
 
 
+def is_sql(text):
+    if text in ("read-only", "readonly"):
+        return Q(state=STATE_READONLY)
+    if text == "approved":
+        return Q(state=STATE_APPROVED)
+    if text == "translated":
+        return Q(state__gte=STATE_TRANSLATED)
+    if text == "untranslated":
+        return Q(state__lt=STATE_TRANSLATED)
+    if text == "pending":
+        return Q(pending=True)
+
+    raise ValueError("Unsupported is lookup: {}".format(text))
+
+
 def query_sql(obj):
     if isinstance(obj, whoosh.query.And):
         return reduce(
@@ -261,6 +281,8 @@ def query_sql(obj):
     if isinstance(obj, whoosh.query.Term):
         if obj.fieldname == "has":
             return has_sql(obj.text)
+        if obj.fieldname == "is":
+            return is_sql(obj.text)
         return field_extra(obj.fieldname, Q(**{field_name(obj.fieldname): obj.text}))
     if isinstance(obj, whoosh.query.DateRange):
         return field_extra(